additionalMetadata

additionalMetadata

additionalMetadata(describes = NULL, metadata = NULL, id = NULL)

Arguments

describes
A pointer to the id attribute for the sub-portion of the resource that is described by this additional metadata. See describes()

metadata
This element contains the additional metadata that is to be included in the document. The content of this field can be any well-formed XML fragment. See metadata()

id
A unique identifier for this additional metadata that can be used to reference it elsewhere. This is a formal field in that it is an error to provide a value for the id attribute that is not unique within the document's set of id attributes. This is designed to allow other portions of the metadata to reference this section formally.

Value

a additionalMetadata list object