imageDescription

imageDescription

imageDescription(illuminationElevationAngle = NULL,
  illuminationAzimuthAngle = NULL, imageOrientationAngle = NULL,
  imagingCondition = NULL, imageQualityCode = NULL,
  cloudCoverPercentage = NULL, preProcessingTypeCode = NULL,
  compressionGenerationQuality = NULL, triangulationIndicator = NULL,
  radiometricDataAvailability = NULL,
  cameraCalibrationInformationAvailability = NULL,
  filmDistortionInformationAvailability = NULL,
  lensDistortionInformationAvailability = NULL, bandDescription = NULL)

Arguments

illuminationElevationAngle
Illumination elevation measured in degrees clockwise from the target plane at intersection of the optical line of sight with the earth's surface.

illuminationAzimuthAngle
Illumination azimuth measured in degrees clockwise from true north at the time the image is taken.

imageOrientationAngle
Angle from the first row of the image to true north in degrees, clockwise.

imagingCondition
Code which indicates conditions which affect the quality of the image.

imageQualityCode
Specifies the image quality.

cloudCoverPercentage
Area of the dataset obscured by clouds, expressed as a percentage of the spatial extent.

preProcessingTypeCode
Image distributor's code that identifies the level of radiometric and geometric processing applied against the image.

compressionGenerationQuality
Counts the number of lossy compression cycles performed on the image.

triangulationIndicator
Code which indicates whether or not triangulation has been performed upon the image.

radiometricDataAvailability
Code which indicates whether or not Standard Radiometric Product data is available.

cameraCalibrationInformationAvailability
Code which indicates whether or not constants are available which allow for camera calibration corrections.

filmDistortionInformationAvailability
Code which indicates whether or not Calibration Reseau information is available

lensDistortionInformationAvailability
Code which indicates whether or not lens aberration correction information is available

bandDescription
Description of the wavelength domain that the sensor operates in.

inheritParams
node_template

Value

a imageDescription list object