protocol

protocol

protocol(alternateIdentifier = NULL, shortName = NULL, title = NULL,
  creator = NULL, metadataProvider = NULL, associatedParty = NULL,
  pubDate = NULL, language = NULL, series = NULL, abstract = NULL,
  keywordSet = NULL, additionalInfo = NULL,
  intellectualRights = NULL, licensed = NULL, distribution = NULL,
  coverage = NULL, annotation = NULL, proceduralStep = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)

Arguments

alternateIdentifier
A secondary identifier for this entity

shortName
A short name that describes the resource, sometimes a filename.

title
A brief description of the resource, providing enough detail to differentiate it from other similar resources.

creator
The people or organizations who created this resource.

metadataProvider
The people or organizations who created provided documentation and other metadata for this resource.

associatedParty
Other people or organizations who should be associated with this resource.

pubDate
The publication date of the resource.

language
The language in which the resource is written.

series
The series from which the resource came.

abstract
A brief overview of the resource.

keywordSet
Keyword information that describes the resource.

additionalInfo
Any extra information pertitent to the resource.

intellectualRights
Intellectual property rights regarding usage and licensing of this resource.

licensed
Information identifying a well-known license for the metadata and data

distribution
Information on how the resource is distributed online and offline

coverage
Extent of the coverage of the resource.

annotation
A precisely-defined semantic statement about this resource.

proceduralStep
(def not found)

references
The id of another element in this EML document to be used to here in this context.

inheritParams
node_template

Value

a protocol list object