relatedProject

relatedProject

relatedProject(title = NULL, personnel = NULL, abstract = NULL,
  funding = NULL, award = NULL, studyAreaDescription = NULL,
  designDescription = NULL, relatedProject = NULL, references = NULL,
  id = NULL, system = NULL, scope = NULL)

Arguments

title
Title of the project.

personnel
Contact and role information for people involved in the research project.

abstract
Project Abstract.

funding
Funding information.

award
Award information.

studyAreaDescription
Description of the physical area associated with the research project, potentially including coverage, climate, geology, disturbances, etc.

designDescription
Description of the design of the research project

relatedProject
This field is a recursive link to another project.

references
The id of another element in this EML document to be used to here in this context.

inheritParams
node_template

Value

a relatedProject list object