spatialRaster

spatialRaster

spatialRaster(alternateIdentifier = NULL, entityName = NULL,
  entityDescription = NULL, physical = NULL, coverage = NULL,
  methods = NULL, additionalInfo = NULL, annotation = NULL,
  attributeList = NULL, constraint = NULL, spatialReference = NULL,
  georeferenceInfo = NULL, horizontalAccuracy = NULL,
  verticalAccuracy = NULL, cellSizeXDirection = NULL,
  cellSizeYDirection = NULL, numberOfBands = NULL,
  rasterOrigin = NULL, rows = NULL, columns = NULL,
  verticals = NULL, cellGeometry = NULL, toneGradation = NULL,
  scaleFactor = NULL, offset = NULL, imageDescription = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)

Arguments

alternateIdentifier
A secondary identifier for this entity See alternateIdentifier()

entityName
The name of the entity name See entityName()

entityDescription
General description of the entity and its contents See entityDescription()

physical
Description of the physical format of the entity See physical()

coverage
Information on the geographic, spatial and temporal coverages used in this entity. See coverage()

additionalInfo
Any extra information pertitent to the entity. See additionalInfo()

annotation
A precisely-defined semantic statement about this entity. See annotation()

attributeList
The list of attributes associated with this entity. See attributeList()

constraint
Description of any relational constraints on this entity. See constraint()

spatialReference
The means by which positional information in this entity is related to points on the earth's surface. See spatialReference()

georeferenceInfo
Information on how to position the grid within the coordinate system defined in the spatial reference module. See georeferenceInfo()

horizontalAccuracy
The accuracy of horizontal locational measurements within the data. See horizontalAccuracy()

verticalAccuracy
The accuracy of vertical locational measurements within the data. See verticalAccuracy()

cellSizeXDirection
The width of the cell in the x direction. See cellSizeXDirection()

cellSizeYDirection
The width of the cell in the x direction. See cellSizeYDirection()

numberOfBands
The number of bands in the image. See numberOfBands()

rasterOrigin
The corner location of the pixel having the minimum x and y values See rasterOrigin()

rows
Maximum number of raster objects along the ordinate (y) axis See rows()

columns
Maximum number of raster objects along the abscissa (x) axis See columns()

verticals
Maximum number of raster objects along the vertical (z) axis See verticals()

cellGeometry
Geometric representation of the cell's content. See cellGeometry()

toneGradation
Number of colors present in the image. See toneGradation()

scaleFactor
The value used for scaling the source raster. See scaleFactor()

offset
The offset value used for scaling the source raster. See offset()

imageDescription
Detailed information for data derived from image sensors. See imageDescription()

references
The id of another element in this EML document to be used to here in this context. See references()

id
A unique identifier for this additional metadata that can be used to reference it elsewhere. This is a formal field in that it is an error to provide a value for the id attribute that is not unique within the document's set of id attributes. This is designed to allow other portions of the metadata to reference this section formally.

system
The data management system within which an identifier is in scope and therefore unique. This is typically a URL (Uniform Resource Locator) that indicates a data management system. All identifiers that share a system must be unique. In other words, if the same identifier is used in two locations with identical systems, then by definition the objects at which they point are in fact the same object.

scope
The scope of the identifier. Scope is generally set to either "system", meaning that it is scoped according to the "system" attribute, or "document" if it is only to be in scope within this single document instance. In this particular use of scope, it is FIXED to be "system" because the packageId is required and always has the scope of the required "system".

Value

a spatialRaster list object