usageCitation

usageCitation

usageCitation(alternateIdentifier = NULL, shortName = NULL,
  title = NULL, creator = NULL, metadataProvider = NULL,
  associatedParty = NULL, pubDate = NULL, language = NULL,
  series = NULL, abstract = NULL, keywordSet = NULL,
  additionalInfo = NULL, intellectualRights = NULL, licensed = NULL,
  distribution = NULL, coverage = NULL, annotation = NULL,
  contact = NULL, article = NULL, book = NULL, chapter = NULL,
  editedBook = NULL, manuscript = NULL, report = NULL,
  thesis = NULL, conferenceProceedings = NULL,
  personalCommunication = NULL, map = NULL, generic = NULL,
  audioVisual = NULL, presentation = NULL, bibtex = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)

Arguments

alternateIdentifier
A secondary identifier for this entitySee alternateIdentifier()

shortName
A short name that describes the resource, sometimes a filename.See shortName()

title
A brief description of the resource, providing enough detail to differentiate it from other similar resources.See title()

creator
The people or organizations who created this resource.See creator()

metadataProvider
The people or organizations who created provided documentation and other metadata for this resource.See metadataProvider()

associatedParty
Other people or organizations who should be associated with this resource.See associatedParty()

pubDate
The publication date of the resource.See pubDate()

language
The language in which the resource is written.See language()

series
The series from which the resource came.See series()

abstract
A brief overview of the resource.See abstract()

keywordSet
Keyword information that describes the resource.See keywordSet()

additionalInfo
Any extra information pertitent to the resource.See additionalInfo()

intellectualRights
Intellectual property rights regarding usage and licensing of this resource.See intellectualRights()

licensed
Information identifying a well-known license for the metadata and dataSee licensed()

distribution
Information on how the resource is distributed online and offlineSee distribution()

coverage
Extent of the coverage of the resource.See coverage()

annotation
A precisely-defined semantic statement about this resource.See annotation()

contact
An optional contact individual for this citationSee contact()

article
Information for full citation of a journal article or other periodical article.See article()

book
Information for full citation of a bookSee book()

chapter
Information for citation of a chapter in a bookSee chapter()

editedBook
Information for full citation of an edited bookSee editedBook()

manuscript
Information about an unpublished manuscriptSee manuscript()

report
Information about a report published by an institution.See report()

thesis
Information about a thesis that has been written as part of a degree requirement and is frequently published in small numbers by the degree awarding institution.See thesis()

conferenceProceedings
The published notes, papers, presentations etc... of a conference.See conferenceProceedings()

personalCommunication
A personal communication that has been saved online or as hard-copy.See personalCommunication()

map
This element describes the map that is being cited or cataloged.See map()

generic
This reference type was created for references that do not fit in to the other existing reference typesSee generic()

audioVisual
This reference type is meant to cover all forms of audio and visual media.See audioVisual()

presentation
An unpublished presentation from a conference, workshop, workgroup, symposium etc.See presentation()

bibtex
Citation in Bibtex format.See bibtex()

references
The id of another element in this EML document to be used to here in this context.See references()

id
A unique identifier for this additional metadata that can be used to reference it elsewhere. This is a formal field in that it is an error to provide a value for the id attribute that is not unique within the document's set of id attributes. This is designed to allow other portions of the metadata to reference this section formally.

system
The data management system within which an identifier is in scope and therefore unique. This is typically a URL (Uniform Resource Locator) that indicates a data management system. All identifiers that share a system must be unique. In other words, if the same identifier is used in two locations with identical systems, then by definition the objects at which they point are in fact the same object.

scope
The scope of the identifier. Scope is generally set to either "system", meaning that it is scoped according to the "system" attribute, or "document" if it is only to be in scope within this single document instance. In this particular use of scope, it is FIXED to be "system" because the packageId is required and always has the scope of the required "system".

Value

a usageCitation list object